+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `PseudoLandmarkDataset` class.
+- `MathematicalLandmarkDataset` class.
 
 ## [1.0.0] - 2025-10-14